Make Searchable specific to pages
I initially extracted a generic Searchable module but now that I’ve had
a chance to see what it would look like, I’m not comfortable with sql
interpolation, even if it‘s for portions of the sql strings that are not
user facing.
